- Add ajax form support in Shortcut

+++ web/frontend/shortcut.ts
@@ -4,20 +4,22 @@ const PREFIX = 'sc';
 const ID = 'id';
 const TEMPLATE_ID = 'template-id';
 const TYPE = 'type';
-// const BIND = 'bind';
-// const BIND_ACTION = 'bind-action';
 const METHOD = 'method';
-const SUCCESS_MESSAGE = 'success-message';
+const FORM_ON_ERROR = 'form-onerror';
+const FORM_ON_SUCCESS = 'form-onsuccess';
 const HANDLER = 'handler';
 
 type Handle = (e: Event, currentCustomElement: CustomElement, is: InternalState) => void;
+type FormMethod = (data: any) => void;
 type InternalState = {
+  formMethods: Map<string, FormMethod>,
   elements: Map<string, CustomElement>;
   handlers: Map<string, Handle>;
   state: Map<string, any>;
   getByType(type: string): CustomElement[];
 }
 const store: InternalState = {
+  formMethods: new Map(),
   elements: new Map(),
   handlers: new Map(),
   state: new Map(),
@@ -114,8 +116,7 @@ export class CustomElement {
     }
 
     const hasAllAttrs = this.ref.getAttribute(`${createDataName(METHOD)}`) &&
-      this.ref.getAttribute('action') &&
-      this.ref.getAttribute(`${createDataName(SUCCESS_MESSAGE)}`);
+      this.ref.getAttribute('action');
 
     if (this.ref instanceof HTMLFormElement && hasAllAttrs) {
       this.overrideSubmit(this.ref);
@@ -130,7 +131,9 @@ export class CustomElement {
 
       const action = el.getAttribute('action');
       const method = el.getAttribute(`${createDataName(METHOD)}`);
-      const successMessage = el.getAttribute(`${createDataName(SUCCESS_MESSAGE)}`);
+      const formOnSuccess = el.getAttribute(`${createDataName(FORM_ON_SUCCESS)}`);
+      const formOnError = el.getAttribute(`${createDataName(FORM_ON_ERROR)}`);
+
       if (!method || !action) {
         console.warn(`${this.overrideSubmit.name}: action or method is missing, skipping`);
         return;
@@ -139,6 +142,7 @@ export class CustomElement {
       const methods = ['POST', 'PUT', 'DELETE', 'PATCH'];
       const upperCasedMethod = method.toUpperCase();
       if (!methods.includes(upperCasedMethod)) {
+        console.warn(`${this.overrideSubmit.name}: method ${method} is not supported, skipping`);
         return;
       }
 
@@ -159,40 +163,52 @@ export class CustomElement {
         }
       }
 
-      const res = await fetch(action, {
-        method: upperCasedMethod,
-        body: JSON.stringify(body),
-        credentials: 'include',
-        headers: {
-          'Content-Type': 'application/json'
+      try {
+        const res = await fetch(action, {
+          method: upperCasedMethod,
+          body: JSON.stringify(body),
+          credentials: 'include',
+          headers: {
+            'Content-Type': 'application/json'
+          }
+        });
+
+        if (!res.ok) {
+          const message = await res.text();
+          if (!formOnError) {
+            console.error('error: ' + message);
+            return;
+          }
+
+          const errorFn = store.formMethods.get(formOnError);
+          if (!errorFn) {
+            return;
+          }
+          errorFn(message);
+          return;
         }
-      });
 
-      if (!res.ok) {
-        const message = await res.text();
-        console.error('error: ' + message);
-        return;
-      }
-
-      const contentType = res.headers.get('Content-Type');
-      if (contentType && contentType?.length > 0) {
-        location.href = '/';
-      }
-
-      if (successMessage) {
-        const successElm = document.createElement('p');
-        successElm.innerText = successMessage;
-        el.appendChild(successElm);
-        setTimeout(() => {
-          successElm.remove();
-        }, 2000);
+        if (formOnSuccess) {
+          const json = await res.json();
+          const successFn = store.formMethods.get(formOnSuccess);
+          if (!successFn) {
+            throw new Error(`${this.overrideSubmit.name}: success handler ${formOnSuccess} not found`);
+          }
+          successFn(json);
+        }
+      } catch (e) {
+        const error = e as Error;
+        console.error(error.message);
       }
     };
     el.addEventListener('submit', handler);
     this.handlers.push(handler as unknown as EventListener);
   };
 
-  insertTemplateInto = (target: CustomElement | Element, clearBeforeInsert?: boolean) => {
+  insertTemplateInto = (target: CustomElement | Element, options?: {
+    clearBeforeInsert?: boolean
+    className?: string
+  }) => {
     const intoTarget = target instanceof CustomElement ? target.ref : target;
 
     if (!this.isTemplate) {
@@ -207,9 +223,12 @@ export class CustomElement {
     const clone = template.content.cloneNode(true) as DocumentFragment;
     wrapper.appendChild(clone);
     wrapper.setAttribute(createDataName(TEMPLATE_ID), this.id);
+    if (options && options.className) {
+      wrapper.classList.add(options.className);
+    }
     this.templateWrapperRef = wrapper;
 
-    if (clearBeforeInsert) {
+    if (options && options.clearBeforeInsert) {
       // TODO test this for normal Element, this might throw error
       deleteAllFromTarget(intoTarget);
       intoTarget.innerHTML = '';
@@ -221,7 +240,7 @@ export class CustomElement {
 
   remove() {
     const ref = this.isTemplate ? this.templateWrapperRef : this.ref;
-    if(ref){
+    if (ref) {
       this.handlers.forEach(handle => {
         if (this.handleEvent) {
           ref.removeEventListener(this.handleEvent, handle);
@@ -232,7 +251,6 @@ export class CustomElement {
     }
   }
 }
-
 
 export const addFromTarget = (target: Element) => {
   const elements = scanElements(target);
@@ -278,6 +296,14 @@ export const addHandler = (handlerName: string, handle: Handle) => {
   }
   store.handlers.set(handlerName, handle);
   syncHandle();
+};
+
+export const addFormMethod = (methodName: string, method: FormMethod) => {
+  const getMethod = store.formMethods.get(methodName);
+  if (getMethod) {
+    throw new Error(`${addFormMethod.name}: method ${methodName} already exists`);
+  }
+  store.formMethods.set(methodName, method);
 };
 
 export const getCookie = (name: string) => {
